Creates KeyFinder module and moves methods from Cracker
